+++ test.js
@@ -5,7 +5,7 @@ const ONE_SEC = 1000;
 logger.log('Hello, loggy');
 logger.warn('Deprecated');
 logger.info(new Date());
-logger.error('Oops');
+logger.error('$PATH');
 
 setTimeout(() => {
   logger.error(new TypeError('undefined is not a function'));
